+++ packages/ui/src/components/ProfileSettings/ProfilePanel.tsx
@@ -1,5 +1,5 @@
 'use client'
-import { FC, PropsWithChildren } from 'react'
+import { Children, FC, isValidElement, PropsWithChildren, useEffect, useMemo, useRef, useState } from 'react'
 
 import {
   Box,
@@ -14,6 +14,7 @@ import {
   VStack,
   useBreakpointValue,
 } from '@chakra-ui/react'
+import { useRouter } from 'next/router'
 import { useTranslation } from 'next-i18next'
 import { FaPaintBrush, FaUserCircle } from 'react-icons/fa'
 import { FaBlog, FaKey } from 'react-icons/fa6'
@@ -36,18 +37,36 @@ type ProfilePanelProps = PropsWithChildren<{
   showArts?: boolean
 }>
 
-const CustomTab = (props: TabProps) => (
-  <Tab
-    borderWidth={1}
-    rounded={'md'}
-    borderColor={'transparent'}
-    _selected={{ borderColor: 'primary.500', color: 'primary.500' }}
-    fontWeight={600}
-    justifyContent={{ base: 'center', lg: 'start' }}
-    w={'full'}
-    {...props}
-  />
-)
+
+const CustomTab: FC<{ title: string } & TabProps> = ({
+  title,
+  ...props
+}) => {
+  return (
+    <Tab
+      borderWidth={1}
+      rounded={'md'}
+      title={title}
+      borderColor={'transparent'}
+      _selected={{ borderColor: 'primary.500', color: 'primary.500' }}
+      fontWeight={600}
+      justifyContent={{ base: 'center', lg: 'start' }}
+      w={'full'}
+      {...props}
+    />
+  )
+}
+
+const findTabIndexByTitle = (tabList: HTMLDivElement | null, title: string): number => {
+  if (!tabList) return 0;
+
+  const tabs = Array.from(tabList.children);
+  const index = tabs.findIndex(
+    (child) => (child as unknown as { title: string }).title === title
+  );
+
+  return index === -1 ? 0 : index;
+};
 
 export const ProfilePanel: FC<ProfilePanelProps> = ({
   children,
@@ -60,6 +79,11 @@ export const ProfilePanel: FC<ProfilePanelProps> = ({
   })
   const { t } = useTranslation()
 
+  const router = useRouter()
+  const activeTab = router.query.tab as string ?? "profile"
+  const tabListRef = useRef<HTMLDivElement>(null)
+  const [activeTabIndex, setActiveTabIndex] = useState(0)
+
   const isBlogsVisible =
     (site === 'dashboard' || site === 'foundation') &&
     (user?.roles.includes('admin') || user?.roles.includes('author'))
@@ -67,6 +91,20 @@ export const ProfilePanel: FC<ProfilePanelProps> = ({
   const isCoursePaymentVisible =
     (site === 'foundation' || site === 'dashboard') &&
     process.env.NODE_ENV === 'development'
+
+  useEffect(() => {
+    const index = findTabIndexByTitle(tabListRef.current, activeTab);
+    setActiveTabIndex(index);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [activeTab, tabListRef.current]);
+
+  const updateSelectedTab = (index: number) => {
+    setActiveTabIndex(index)
+    const tabList = Array.from(tabListRef.current?.children || []);
+    const title = (tabList[index] as unknown as { title: string }).title
+
+    router.push(`/profile?tab=${title}`)
+  }
 
   if (!user) return <Hero></Hero>
 
@@ -99,6 +137,8 @@ export const ProfilePanel: FC<ProfilePanelProps> = ({
             orientation={orientation}
             border={0}
             colorScheme="primary"
+            index={activeTabIndex}
+            onChange={updateSelectedTab}
             size="lg"
             variant={'unstyled'}
             gap={8}
@@ -109,34 +149,35 @@ export const ProfilePanel: FC<ProfilePanelProps> = ({
               minW={{ base: 'auto', lg: 300 }}
               mb={{ base: 8, lg: 0 }}
               h="max-content"
+              ref={tabListRef}
               overflowX={'auto'}
             >
-              <CustomTab>
+              <CustomTab title={'profile'}>
                 <Box as={FaUserCircle} mr={2} />
                 <Box>{t('profile.tabs.profile')}</Box>
               </CustomTab>
-              <CustomTab>
+              <CustomTab title={'security'}>
                 <Box as={FaKey} mr={2} />
                 <Box>{t('profile.tabs.security')}</Box>
               </CustomTab>
-              <CustomTab>
+              <CustomTab title={'socials'}>
                 <Box as={TbSocial} mr={2} />
                 <Box>{t('profile.tabs.socials')}</Box>
               </CustomTab>
               {isCoursePaymentVisible && (
-                <CustomTab>
+                <CustomTab title={'courses'}>
                   <Box as={MdOutlinePayments} mr={2} />
                   <Box>{t('profile.tabs.courses')}</Box>
                 </CustomTab>
               )}
               {showArts && (
-                <CustomTab>
+                <CustomTab title={'arts'}>
                   <Box as={FaPaintBrush} mr={2} />
                   <Box>{t('profile.tabs.arts')}</Box>
                 </CustomTab>
               )}
               {isBlogsVisible && (
-                <CustomTab>
+                <CustomTab title={'blogs'}>
                   <Box as={FaBlog} mr={2} />
                   <Box>{t('profile.tabs.blogs')}</Box>
                 </CustomTab>
